+++ webapp/src/payload/collections/Partner.ts
@@ -40,5 +40,13 @@ export const Partners: CollectionConfig = {
       required: true,
       relationTo: "media",
     },
+    {
+      name: "stared",
+      type: "checkbox",
+      label: "Mettre en avant",
+      admin: {
+        position: "sidebar",
+      },
+    },
   ],
 };