+++ webapp/src/payload/migrations/20240319_124219.ts
@@ -1,0 +1,16 @@
+import { MigrateUpArgs, MigrateDownArgs } from '@payloadcms/db-postgres'
+import { sql } from 'drizzle-orm'
+
+export async function up({ payload }: MigrateUpArgs): Promise<void> {
+await payload.db.drizzle.execute(sql`
+
+ALTER TABLE "partners" ADD COLUMN "stared" boolean;`);
+
+};
+
+export async function down({ payload }: MigrateDownArgs): Promise<void> {
+await payload.db.drizzle.execute(sql`
+
+ALTER TABLE "partners" DROP COLUMN IF EXISTS "stared";`);
+
+};